Fixed an error with GC
It wouldn't mark the result after finishing evaluate_expr, so unless it was set to a symbol it would delete it. This caused lists to not be displayed by say_expr after being created by certain functions. All I did was make it mark the result so now it doesn't delete the result right after.